move SetCurrentContainerImages from pkg/utils to aslan/common/utils
