+++ pkg/microservice/aslan/core/service/service/production_service.go
@@ -30,6 +30,7 @@ import (
 	templaterepo "github.com/koderover/zadig/pkg/microservice/aslan/core/common/repository/mongodb/template"
 	"github.com/koderover/zadig/pkg/microservice/aslan/core/common/service"
 	commonservice "github.com/koderover/zadig/pkg/microservice/aslan/core/common/service"
+	commonutil "github.com/koderover/zadig/pkg/microservice/aslan/core/common/util"
 	"github.com/koderover/zadig/pkg/setting"
 	e "github.com/koderover/zadig/pkg/tool/errors"
 	"github.com/koderover/zadig/pkg/tool/log"
@@ -184,7 +185,7 @@ func ensureProductionServiceTmpl(args *commonmodels.Service, log *zap.SugaredLog
 
 	// since service may contain go-template grammar, errors may occur when parsing as k8s workloads
 	// errors will only be logged here
-	if err := util.SetCurrentContainerImages(args); err != nil {
+	if err := commonutil.SetCurrentContainerImages(args); err != nil {
 		log.Errorf("failed to ser set container images, err: %s", err)
 	}
 	log.Infof("find %d containers in service %s", len(args.Containers), args.ServiceName)
@@ -226,7 +227,7 @@ func UpdateProductionServiceVariables(args *commonservice.ServiceTmplObject) err
 	currentService.RenderedYaml = util.ReplaceWrapLine(currentService.RenderedYaml)
 	currentService.KubeYamls = util.SplitYaml(currentService.RenderedYaml)
 	oldContainers := currentService.Containers
-	if err := util.SetCurrentContainerImages(currentService); err != nil {
+	if err := commonutil.SetCurrentContainerImages(currentService); err != nil {
 		log.Errorf("failed to ser set container images, err: %s", err)
 		//return err
 	} else if containersChanged(oldContainers, currentService.Containers) {
